EHN: add cli
