--- bin/pfla
@@ -1,2 +1,0 @@
-#!/usr/bin/env python
-from pfla import __init__
